The intent service only needs to handle price changes, everything else goes to the activities
