+++ java/app/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
         <activity
             android:name=".activity.CheckoutActivity"
             android:label="@string/store_name"
+            android:theme="@style/Theme.AppCompat"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
